requirements.txt
torch
torchvision
matplotlib
pillow
numpy
scikit-learn
jupyter
